requirements.txt
flask==2.3.3
cohere==4.30.0
python-dotenv==1.0.0
openai==1.6.1
reportlab==4.0.4
twilio==8.2.2
requests==2.31.0
pypdf2==3.0.1
pytest==7.4.0
markdown==3.4.4
pillow==10.0.0
weasyprint==54.2
html5lib==1.1
cssselect2==0.7.0
cffi>=0.6
pydyf>=0.0.3
fonttools[woff]>=4.0.0
tinycss2>=1.0.0 